+++ source/aspnetcore-tracing-sample/WebApp1/Program.cs
@@ -1,3 +1,5 @@
+using Microsoft.Extensions.Azure;
+
 namespace WebApp1
 {
     public class Program
@@ -11,6 +13,12 @@ namespace WebApp1
 
             builder.Services.AddApplicationInsightsTelemetry();
             builder.Services.AddHttpClient();
+            builder.Services.AddAzureClients(acfbuilder =>
+            {
+                var strconstr = builder.Configuration["STORAGE_CONNECTION_STRING"]!.ToString();
+                acfbuilder.AddBlobServiceClient(strconstr);
+            });
+
 
             var app = builder.Build();
